rename: 글작성  api및 hook을 각 main 폴더로 이동
